pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "octopin"
dynamic = ["version"]
description = "Pin used actions and analyse transitive dependencies of GitHub workflows / actions"
authors = [
    {name = "Thomas Neidhart", email = "thomas.neidhart@eclipse-foundation.org"},
]
requires-python = ">=3.11"

classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python",
    "Topic :: Security",
    "Topic :: Software Development",
    "Typing :: Typed",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: Eclipse Public License 2.0 (EPL-2.0)",
]

license = {text = "EPL-2.0"}
readme = "README.md"

dependencies = [
    "typer >= 0.15",
    "PyYAML >= 6.0",
    "semver >= 3.0",
    "aiohttp >= 3.9",
    "aiohttp-client-cache >= 0.10",
    "aiosqlite >= 0.19",
    "platformdirs >= 4",
]

[project.urls]
Homepage      = "https://github.com/eclipse-csi/octopin"
Documentation = "https://octopin.readthedocs.org"
Repository    = "https://github.com/eclipse-csi/octopin"
Issues        = "https://github.com/eclipse-csi/octopin/issues"
Changelog     = "https://octopin.readthedocs.org/release-notes/"

[project.scripts]
octopin = "octopin.cli:main"

[dependency-groups]
dev = [
    "ruff >= 0.8",
    "mypy >= 1.8",
    "types-PyYAML >= 6.0",
    "pre-commit >= 4",
    "pytest >=8, <9",
    "pytest-asyncio >= 0.24",
    "parameterized  >= 0.9",
    "tox >= 4.22, <5"
]

[tool.pdm]
version = { source = "file", path = "octopin/__init__.py" }
distribution = true

[tool.pdm.build]
source-includes = [
    "octopin/",
    "tests/",
]

[tool.pytest.ini_options]
minversion = "7.2"
testpaths  = ["tests"]

log_cli             = true
log_cli_level       = "INFO"
log_cli_format      = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

asyncio_default_fixture_loop_scope="function"

[tool.mypy]
python_version = "3.11"
strict         = true
exclude        = ["docs"]

[[tool.mypy.overrides]]
module = [
    "parameterized"
]
ignore_missing_imports = true

[tool.ruff]
line-length    = 120
indent-width   = 4
target-version = "py311"

[tool.ruff.lint]
extend-select = [
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ERA", # flake8-eradicate/eradicate
    "I",   # isort
    "N",   # pep8-naming
    "PIE", # flake8-pie
    "S",   # flake8-bandit
    "PGH", # pygrep
    "RUF", # ruff checks
    "SIM", # flake8-simplify
    "T20", # flake8-print
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "UP",  # pyupgrade
]
ignore = [
    "E203",
    "ERA001", # commented out code
    "PGH003", # specific ignore
    "PGH004", # specific ignore
    "SIM102", # nested ifs
    "SIM103", # inline return
    "N818",   # exception names
    "S102",   # we deliberately use exec
    "S105"    # hardcoded passwords (too many false-positives for now)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.tox]
requires = ["tox>=4.22"]
env_list = ["3.11", "3.12", "3.13"]

[tool.tox.env_run_base]
description = "Run tests under {base_python}"
skip_install = true
allowlist_externals = ["uv"]
commands_pre = [["uv", "sync"]]
commands = [["uv", "run", "pytest", "tests/", "--import-mode", "importlib"]]
